Bird-4: The Anti-Gravity Update
Within main: Initialized a table to store keysPressed, added a function to add our table of keys pressed in a frame, added another function to check the global input table input for keys we activate in a frame, and reset the table in love.update function. Within Bird class: updated gravity constant to 20 and added a check to see if 'space' was pressed within the init function for Bird.