Bird-5: The Infinite Pipe Update
Added in the pipe image, the Pipe class which initializes the pipe, updates its scrolling, renders/draws it to the screen and loads the image just once since many will spawn we don't need to initialize it over and over. Also updated main with a spawn timer, pipes, and render/drew it on the screen randomly and infinitely.